Now linking the user binary into the app dir.
Note: I've been going to a lot of trouble to get the scsh VM into the
app directory, but this turns out to have been a waste of time, since
the VM itself uses dynamically linked system libraries. This means
that I can't deploy a static binary to users (They will have to
install Scsh on their systems somehow. Shit.) Therefore, the next step
is to investigate building Scsh as a static binary.